220 uv auth0 authentication (#231)
* updated task creation parameters

* updated satellite source param key

* fixed api tests

* frontend:updated cloud cover param key

* backend:updated cloud cover param key

* feat: added initial data processor proxy module

* feat:frontend: integrated results from data processing

* adjusted records dataframe to consider one mgrs tile id from bounds tile ids

* feat:backend: integrated data extraction for apps component

* fixed earthdata credentials for right user and bbox data extraction call

* updated readme

* updated readme

* updated readme

* updated times to utc times and formatting on frontend

* removed comment on default params

* revert geo utils

* include required packages for model evaluation

* Add model prediction pipeline

* fix: fix linting

* fix: fix linting errors

* fix: fix linting errors

* resolve comments on PR

* Add comment to clarify configs.yaml current usage

* Add model loading and restructuring

* Add class configs

* Add class configs

* refactor: Restructure configs

* fix: regroup configs and params

* revert changes to files

* Address PR comments

* Address PR comments

* Address PR comments

* Address PR comments

* Address PR comments

* Update instageo/data/data_pipeline.py


* Update instageo/model/factory.py


* Update instageo/model/inference_pipeline.py


* Address PR comments

* Address PR comments

* Add carbon tracking for chip inference

* Add api endpoint to load all models available

* Fix: fix model prediction pipeline bugs

* fix docstrings

* Update instageo/model/infer_utils.py


* Update instageo/model/infer_utils.py


* Update instageo/model/infer_utils.py


* integrate model pipeline

* Resolve merge conflicts in biomass scripts and manifest

* add: app integration

* latest working version

* added model results stats in task monitor

* adjusted num params formatting

* fix:remove unwanted files

* fix: fix class mapping and model registry errors

* feat: added func to get complete chips for bbox

* update: update config.env.example

* added formatting for model results on tasks monitor

* fixed tooltip error on tasks monitor

* update: update readme.md

* fixed bug related to segmentation predictions

* bugs fixing and misc improvements

* requirements: added codecarbon

* adjusted config and dockerfile

* updated readme

* updated model registry sync script

* updated readme

* revamped frontend

* added port mapping for redis

* added misc improvements

* fixed api tests

* added port mapping for rqdashboard prod

* adjusted redis port mapping to use env var

* fixed precommit EOF issue

* adjusted rq dashboard for prod

* fix: configure black line length to match flake8 (100 chars)

* feat: enabled tunneling with cloudflared

* fixed typo in api base url

* revert changes to branch

* fix: apply changes lost in rebase process

* removed unnecessary rules from nginx conf prod

* chore: address review comment

* add: uv for package management

* add: uv for package management

* add: fix models not iterable issue

* add: uv for package management

* style: merge backend dockerfile with latest changes
- update python version to 3.10
- update dockerfile to use uv package manager

* fix: fix github actions ci to work with uv

# Conflicts:
#	.github/workflows/tests_and_linters.yaml

* fix: fix github actions ci to work with uv

* fix: fix github actions ci to work with uv

* fix: fix github actions ci to work with uv

* fix: fix github actions ci to work with uv

* fix: fix github actions ci to work with uv

* fix: fix github actions ci to work with uv

* fix: fix github actions ci to work with uv

* upgrade ci 10 python 3.11

* fix: debug ci failure

* fix: fix uv dependencies & ci pipeline

* fix: fix uv dependencies & ci pipeline

* remove unused code snippet

* fix: fix torch version in ci

* fix: fix torch version in ci

* fix: fix torch version in ci

* fix: fix torch version in ci

* fix: fix torch version in ci

* fix: fix torch version in ci

* fix: fix torch version in ci

* fix: fix torch version in ci

* Update tests_and_linters.yaml

* Update tests_and_linters.yaml

* Update tests_and_linters.yaml

* fix: fix torch version in ci

* Update tests_and_linters.yaml

* fix: fix torch gpu in ci

* fix: fix torch gpu in ci

* fix: fix torch gpu in ci

* fix: fix torch gpu in ci

* fix: fix torch gpu in ci

* fix: fix torch gpu in ci

* Update tests_and_linters.yaml
# Conflicts:
#	.github/workflows/tests_and_linters.yaml

* fix: fix torch gpu in ci

* fix: fix torch gpu oom in ci

# Conflicts:
#	.github/workflows/tests_and_linters.yaml
#	pyproject.toml

* fix: fix torch gpu oom in ci

* lint: linting and formatting by pre-commit

* add: add ptflops to dependencies in uv lock

* Update tests_and_linters.yaml

* fix: fix torch gpu oom in ci

* fix: fix torch gpu oom in ci

* fix: fix torch gpu oom in ci

* fix: fix torch gpu oom in ci

* fix: remove requirements files

* update: update ci file

* update: update uv lock and pyproject file

* update: update Dockerfile and .dockerignore

* add: add authentication with auth0

# Conflicts:
#	instageo/new_apps/frontend/package.json
#	instageo/new_apps/frontend/src/components/ControlPanel.js
#	instageo/new_apps/frontend/src/index.js

* fix: linting and formatting

* update: update example config file

# Conflicts:
#	instageo/new_apps/config.env.example

* clean up resolved comments

* giMerge branch 'uv-package-mgt' of github.com:instadeepai/InstaGeo into uv-package-mgt

* fix: fix bug in dockerfile

* fix: check memory requirements for ci

* Update config.env.example

* Update auth0-config.js

* fix: fix console logs to frontend

* update: update frontend env variables

* Add: refactor with a profile menu component

# Conflicts:
#	instageo/new_apps/docker-compose.dev.yml
#	instageo/new_apps/frontend/package.json
#	instageo/new_apps/frontend/src/components/ControlPanel.js

* update pyproject  toml

* Update pyproject.toml


* Update pyproject.toml


* Update pyproject.toml


* refactor and clean up code

* fix: mask_segmentation_map dtype mismatch in test

* update: remove unused package

* fix: fix mask_segmentation_map datatype error

* update: update build backend for package

* doc: update README.md on package installation with uv

* Update README.md

* update: README.md project directory name

* doc: update README.md on models local registry path

* fix: frontend base url in configs

* update: upgrade frontend packages

* update: README.md on using uv to run tasks

* update: revert test changes

* update: improve ci script

* update: improve ci script

* update: improve ci script

* update: improve ci script

* update: improve ci script

* update: improve ci script

* update: improve ci script

* update: improve ci script

* update: improve ci script

* update: improve ci script

* refactor to speed up deployment build

* fix: fix dtype wrap bug occuring when casting -1 to uint8

* ci: test new ci file

* fix: replace uint8 with int8 everywhere

* chore: address review comment

* chore: address review comment

* debug: check ci python path

* fix: fix setup package directories in pyproject.toml

* fix: fix setup package directories in pyproject.toml

* fix: fix setup package directories in pyproject.toml

* fix: fix setup package directories in pyproject.toml

* update precommit file

* fix: fix setup package directories in pyproject.toml

* fix: fix library studs import error

* chore: address review comment

---------

+++ tests/data_tests/test_create_chips.py
@@ -82,7 +82,7 @@ def test_segmentation_map_masking():
     seg_map_path = "tests/data/chip_178_022.mask.tif"
     chip_no_data_value = -9999
     chip = rioxarray.open_rasterio(chip_path)
-    seg_map = rioxarray.open_rasterio(seg_map_path)
+    seg_map = rioxarray.open_rasterio(seg_map_path).astype(chip.dtype)
     seg_map = seg_map.assign_coords(x=chip.x.values, y=chip.y.values)
     seg_map = mask_segmentation_map(chip, seg_map, chip_no_data_value)
     assert seg_map.where(seg_map != NO_DATA_VALUES.SEG_MAP).count().values == 0
